worker_node/requirements.txt
grpcio
grpcio-tools
Pillow
requests
psutil
